refactor(source): extract shared same-origin helper
